move generateNavigation to navigation.js
